+++ setup.sh
@@ -4,9 +4,12 @@ RPI_SETUP_DIR="$( pwd )"
 
 I2S_MODE=
 XMOS_DEVICE=
-
+INSTALL_ATTEMPT_NUM_MAX=10
 # Valid values for XMOS device
-VALID_XMOS_DEVICES="xvf3100 xvf3500 xvf3510 xvf3600-slave xvf3600-master xvf3610-int xvf3610-ua xvf3615-int xvf3615-ua"
+VALID_XMOS_DEVICES="xvf3100 xvf3500 xvf3510-int xvf3510-ua xvf3600-slave xvf3600-master xvf3610-int xvf3610-ua xvf3615-int xvf3615-ua xvf3800-intdev xvf3800-inthost"
+
+PACKAGES_TO_INSTALL="python3-matplotlib python3-numpy libatlas-base-dev audacity libreadline-dev libncurses-dev"
+PACKAGES_TO_INSTALL_ONLY_FOR_UA="libusb-1.0-0-dev libevdev-dev libudev-dev"
 
 usage() {
   local VALID_XMOS_DEVICES_DISPLAY_STRING=
@@ -42,7 +45,7 @@ validate_device() {
   return 1
 }
 
-if [[ $# == 1 ]]; then
+if [[ $# -eq 1 ]]; then
   XMOS_DEVICE=$1
   if ! validate_device $XMOS_DEVICE $VALID_XMOS_DEVICES; then
     echo "error: $XMOS_DEVICE is not a valid device type."
@@ -57,36 +60,47 @@ fi
 
 # Configure device-specific settings
 case $XMOS_DEVICE in
-  xvf3[56]10-ua)
-    UA_MODE=y
+  xvf3510-ua|xvf3610-ua|xvf3615-ua)
+    USB_MODE=y
+    I2S_MODE=slave
     ASOUNDRC_TEMPLATE=$RPI_SETUP_DIR/resources/asoundrc_vf_xvf3510_ua
     ;;
-
-  xvf3[56]10-int)
+  xvf3510-int|xvf3610-int|xvf3615-int)
     I2S_MODE=master
-    DAC_SETUP=y
+    IO_EXP_AND_DAC_SETUP=y
     ASOUNDRC_TEMPLATE=$RPI_SETUP_DIR/resources/asoundrc_vf_xvf3510_int
     ;;
   xvf3500)
     I2S_MODE=slave
     ASOUNDRC_TEMPLATE=$RPI_SETUP_DIR/resources/asoundrc_vf_stereo
     ;;
-  xvf3100)
+  xvf3[01]00)
     I2S_MODE=slave
     ASOUNDRC_TEMPLATE=$RPI_SETUP_DIR/resources/asoundrc_vf
     ;;
   xvf3600-slave)
     I2S_MODE=master
-    DAC_SETUP=y
+    IO_EXP_AND_DAC_SETUP=y
     ASOUNDRC_TEMPLATE=$RPI_SETUP_DIR/resources/asoundrc_vf
     ;;
   xvf3600-master)
     I2S_MODE=slave
-    DAC_SETUP=y
+    IO_EXP_AND_DAC_SETUP=y
+    ASOUNDRC_TEMPLATE=$RPI_SETUP_DIR/resources/asoundrc_vf
+    ;;
+  # Note DAC is not setup for XVF3800 - the setup script takes an arg which will conditionally do this
+  xvf3800-intdev)
+    I2S_MODE=master
+    IO_EXP_AND_DAC_SETUP=y
+    ASOUNDRC_TEMPLATE=$RPI_SETUP_DIR/resources/asoundrc_vf
+    ;;
+  xvf3800-inthost)
+    I2S_MODE=slave
+    IO_EXP_AND_DAC_SETUP=y
     ASOUNDRC_TEMPLATE=$RPI_SETUP_DIR/resources/asoundrc_vf
     ;;
   *)
-    echo error: unknown XMOS device type $XMOS_DEVICE.
+    echo Error: unknown XMOS device type $XMOS_DEVICE.
     exit 1
   ;;
 esac
@@ -116,21 +130,32 @@ sudo raspi-config nonint do_spi 0
 # been tested and verified.
 KERNEL_HEADERS_PACKAGE=raspberrypi-kernel-headers
 if ! dpkg -s $KERNEL_HEADERS_PACKAGE &> /dev/null; then
-    echo "Installing Raspberry Pi kernel headers"
-    sudo apt-get install -y $KERNEL_HEADERS_PACKAGE
+  echo "Installing Raspberry Pi kernel headers"
+  sudo apt-get install -y $KERNEL_HEADERS_PACKAGE
 fi
 
-echo "Installing the Python3 packages and related libs"
-sudo apt-get install -y python3-matplotlib
-sudo apt-get install -y python3-numpy
-sudo apt-get install -y libatlas-base-dev
-
 echo  "Installing necessary packages for dev kit"
-sudo apt-get install -y libusb-1.0-0-dev libreadline-dev libncurses-dev libevdev-dev libudev-dev
-
+packages=$PACKAGES_TO_INSTALL
+# Add packages for UA mode
+if [[ -n "$USB_MODE" ]]; then
+  packages="$packages $PACKAGES_TO_INSTALL_ONLY_FOR_UA"
+fi
+for package in $packages; do
+  installed=0
+  attempt_num=0
+  while [ $installed -eq 0 ]; do
+    attempt_num=$((attempt_num+1))
+    sudo apt-get install -y $package && installed=1
+    if [[ $attempt_num -gt $INSTALL_ATTEMPT_NUM_MAX ]]; then
+      echo "Error: installation of package $package failed after $attempt_num attempts"
+      echo "Please retry installation procedure."
+      exit 1
+    fi
+  done
+done
 # Build I2S kernel module
 PI_MODEL=$(cat /proc/device-tree/model | awk '{print $3}')
-if [[ $PI_MODEL = 4 ]]; then
+if [[ $PI_MODEL -eq 4 ]]; then
   I2S_MODULE_CFLAGS="-DRPI_4B"
 fi
 
@@ -147,7 +172,7 @@ if [[ -n "$I2S_MODE" ]]; then
       # no flags needed for I2S slave compilation
       ;;
     *)
-      echo error: I2S mode not known for XMOS device $XMOS_DEVICE.
+      echo Error: I2S mode not known for XMOS device $XMOS_DEVICE.
       exit 1
     ;;
   esac
@@ -169,7 +194,8 @@ fi
 popd > /dev/null
 
 # Copy the udev rules files if device is UA
-if [[ -n "$UA_MODE" ]]; then
+if [[ -n "$USB_MODE" ]]; then
+  echo "Add UDEV rules for XMOS devices"
   sudo cp $RPI_SETUP_DIR/resources/99-xmos.rules /etc/udev/rules.d/
 fi
 
@@ -184,82 +210,100 @@ fi
 
 # Check XMOS device for asoundrc selection.
 if [[ -z "$ASOUNDRC_TEMPLATE" ]]; then
-  echo error: sound card config not known for XMOS device $XMOS_DEVICE.
+  echo Error: sound card config not known for XMOS device $XMOS_DEVICE.
   exit 1
 fi
-cp $ASOUNDRC_TEMPLATE ~/.asoundrc
-
-# Make the asoundrc file read-only otherwise lxpanel rewrites it
-# as it doesn't support anything but a hardware type device
-chmod a-w ~/.asoundrc
 
 # Apply changes
 sudo /etc/init.d/alsa-utils restart
 
 if [[ -n "$I2S_MODE" ]]; then
-    # Create the script to run after each reboot and make the soundcard available
-    i2s_driver_script=$RPI_SETUP_DIR/resources/load_i2s_driver.sh
-    rm -f $i2s_driver_script
+  # Create the script to run after each reboot and make the soundcard available
+  i2s_driver_script=$RPI_SETUP_DIR/resources/load_i2s_${I2S_MODE}_driver.sh
+  rm -f $i2s_driver_script
 
-    # Sometimes with Buster on RPi3 the SYNC bit in the I2S_CS_A_REG register is not set before the drivers are loaded
-    # According to section 8.8 of https://cs140e.sergio.bz/docs/BCM2837-ARM-Peripherals.pdf
-    # this bit is set after 2 PCM clocks have occurred.
-    # To avoid this issue we add a 1-second delay before the drivers are loaded
-    echo "sleep 1"  >> $i2s_driver_script
+  # Sometimes with Buster on RPi3 the SYNC bit in the I2S_CS_A_REG register is not set before the drivers are loaded
+  # According to section 8.8 of https://cs140e.sergio.bz/docs/BCM2837-ARM-Peripherals.pdf
+  # this bit is set after 2 PCM clocks have occurred.
+  # To avoid this issue we add a 1-second delay before the drivers are loaded
+  echo "sleep 1"  >> $i2s_driver_script
 
-    I2S_NAME=i2s_$I2S_MODE
-    I2S_MODULE=$RPI_SETUP_DIR/loader/$I2S_NAME/${I2S_NAME}_loader.ko
-    echo "sudo insmod $I2S_MODULE"                            >> $i2s_driver_script
+  I2S_NAME=i2s_$I2S_MODE
+  I2S_MODULE=$RPI_SETUP_DIR/loader/$I2S_NAME/${I2S_NAME}_loader.ko
+  echo "sudo insmod $I2S_MODULE"                            >> $i2s_driver_script
 
-    echo "# Run Alsa at startup so that alsamixer configures" >> $i2s_driver_script	
-    echo "arecord -d 1 > /dev/null 2>&1"                      >> $i2s_driver_script	
-    echo "aplay dummy > /dev/null 2>&1"                       >> $i2s_driver_script
+  echo "# Run Alsa at startup so that alsamixer configures" >> $i2s_driver_script	
+  echo "arecord -d 1 > /dev/null 2>&1"                      >> $i2s_driver_script	
+  echo "aplay dummy > /dev/null 2>&1"                       >> $i2s_driver_script
+
+  if [[ "$I2S_MODE" = "master" ]]; then
+    echo "# Preconfigure i2s clocks to 48kHz"               >> $i2s_driver_script
+    # wait a bit as it doesn't work otherwise, this is probably caused
+    # by the same process that is deleting .asoundrc
+    echo "sleep 15"                                         >> $i2s_driver_script
+    echo "arecord -Dhw:sndrpisimplecar,0 -c2 -fS32_LE -r48000 -s1 /dev/null" >> $i2s_driver_script
+    echo "sudo $RPI_SETUP_DIR/resources/clk_dac_setup/setup_bclk" >> $i2s_driver_script
+  fi
 fi
 
-if [[ -n "$DAC_SETUP" ]]; then
+if [[ -n "$IO_EXP_AND_DAC_SETUP" ]]; then
   pushd $RPI_SETUP_DIR/resources/clk_dac_setup/ > /dev/null
   make
   popd > /dev/null
   dac_and_clks_script=$RPI_SETUP_DIR/resources/init_dac_and_clks.sh
   rm -f $dac_and_clks_script
   # Configure the clocks only if RaspberryPi is configured as I2S master
-  if [[ "$I2S_MODE" == "master" ]]; then
+  if [[ "$I2S_MODE" = "master" ]]; then
     echo "sudo $RPI_SETUP_DIR/resources/clk_dac_setup/setup_mclk"                  >> $dac_and_clks_script
     echo "sudo $RPI_SETUP_DIR/resources/clk_dac_setup/setup_bclk"                  >> $dac_and_clks_script
   fi
   # Note that only the substring xvfXXXX from $XMOS_DEVICE is used in the lines below
-  echo "python $RPI_SETUP_DIR/resources/clk_dac_setup/setup_dac.py $(echo $XMOS_DEVICE | cut -c1-7)" >> $dac_and_clks_script
+  echo "python $RPI_SETUP_DIR/resources/clk_dac_setup/setup_io_exp_and_dac.py $(echo $XMOS_DEVICE | cut -c1-7)" >> $dac_and_clks_script
   echo "python $RPI_SETUP_DIR/resources/clk_dac_setup/reset_xvf.py $(echo $XMOS_DEVICE | cut -c1-7)" >> $dac_and_clks_script
 fi
 
-sudo apt-get install -y audacity
-if [[ -n "$DAC_SETUP" ]]; then
+if [[ -n "$IO_EXP_AND_DAC_SETUP" ]]; then
   audacity_script=$RPI_SETUP_DIR/resources/run_audacity.sh
   rm -f $audacity_script
   echo "#!/usr/bin/env bash" >> $audacity_script
   echo "/usr/bin/audacity &" >> $audacity_script
   echo "sleep 5" >> $audacity_script
-  if [[ "$I2S_MODE" == "master" ]]; then
+  if [[ "$I2S_MODE" = "master" ]]; then
     echo "sudo $RPI_SETUP_DIR/resources/clk_dac_setup/setup_bclk >> /dev/null" >> $audacity_script
   fi
   sudo chmod +x $audacity_script
   sudo mv $audacity_script /usr/local/bin/audacity
 fi
 
+# Regenerate crontab file with new commands
+crontab_file=$RPI_SETUP_DIR/resources/crontab
+if [ -n "$USB_MODE" ]; then
+    crontab_file="${crontab_file}_usb"
+elif [ -n "$I2S_MODE" ]; then
+    crontab_file="${crontab_file}_i2s_${I2S_MODE}"
+fi
+
+rm -f $crontab_file
+
 # Setup the crontab to restart I2S at reboot
-if [ -n "$I2S_MODE" ] || [ -n "$DAC_SETUP" ]; then
-  rm -f $RPI_SETUP_DIR/resources/crontab
-
+if [ -n "$I2S_MODE" ] || [ -n "$IO_EXP_AND_DAC_SETUP" ]; then
   if [[ -n "$I2S_MODE" ]]; then
-    echo "@reboot sh $i2s_driver_script" >> $RPI_SETUP_DIR/resources/crontab
+    echo "@reboot sh $i2s_driver_script" >> $crontab_file
   fi
-
-  if [[ -n "$DAC_SETUP" ]]; then
-    echo "@reboot sh $dac_and_clks_script" >> $RPI_SETUP_DIR/resources/crontab
+  if [[ -n "$IO_EXP_AND_DAC_SETUP" ]]; then
+    echo "@reboot sh $dac_and_clks_script" >> $crontab_file
   fi
-  crontab $RPI_SETUP_DIR/resources/crontab
 popd > /dev/null
 fi
+
+# Setup the crontab to copy the .asoundrc file at reboot
+# Delay the action by 10 seconds to allow the host to boot up
+# This is needed to address the known issue in Raspian Buster:
+# https://forums.raspberrypi.com/viewtopic.php?t=295008
+echo "@reboot sleep 15 && cp $ASOUNDRC_TEMPLATE ~/.asoundrc" >> $crontab_file
+
+# Update crontab
+crontab $crontab_file
 
 echo "To enable all interfaces, this Raspberry Pi must be rebooted."